requirements.txt
flask>=3.0
flask-login>=0.6
requests>=2.32
beautifulsoup4>=4.12
python-dotenv>=1.0
APScheduler>=3.10
googlesearch-python>=1.2
lxml>=5.0
gunicorn>=21.2
psycopg2-binary>=2.9
